+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## 0.26.0 (Unreleased)
 
 ENHANCEMENTS:
+* Add automatic certificate renewal capability to certificate shared service
 * Create CODEOWNERS file with repository maintainers
 * Change Guacamole VM OS disk defaults to Standard SSD ([#4621](https://github.com/microsoft/AzureTRE/issues/4621))
 * Add additional Databricks, Microsoft & Python domains to allowed-dns.json ([#4636](https://github.com/microsoft/AzureTRE/pull/4636))
@@ -19,6 +20,9 @@ BUG FIXES:
 * Add missing image_gallery_id parameter to porter.yaml for guacamole export review vm ([#4678](https://github.com/microsoft/AzureTRE/pull/4678)]
 * Update Starlette and FastAPI versions ([#4683](https://github.com/microsoft/AzureTRE/pull/4683))
 * Fix recreation of route table associations ([#4666](https://github.com/microsoft/AzureTRE/issues/4666))
+
+COMPONENTS:
+* Certificate shared service `tre-shared-service-certs` to version 0.8.0
 
 ## 0.25.0 (July 18, 2025)
 **IMPORTANT**: